vwap-trader/requirements.txt
numpy>=1.26.0
pandas>=2.2.0
tvdatafeed-enhanced>=2.2.1
python-dotenv>=1.2.0
pytz>=2025.2
requests>=2.32.0
websockets>=15.0.0
python-dateutil>=2.9.0
tzdata>=2025.3
